:recycle: refactor: Refactor HomePage Package
