+++ src/App.jsx
@@ -1,27 +1,31 @@
+import React from 'react';
 import { BrowserRouter as Router, Route, Routes, Link } from 'react-router-dom';
 import { MantineProvider } from '@mantine/core';
 import './App.css';
-import HomePage from './pages/HomePage';
-import SearchPage from './pages/SearchPage';
-import ProductDetailPage from './pages/ProductDetailPage';
+import HomePage from './homepage/pages/HomePage';
+import SearchPage from './homepage/pages/SearchPage';
+import ProductDetailPage from './homepage/pages/ProductDetailPage';
+import PaymentPage from './homepage/pages/PaymentPage'; // PaymentPage 컴포넌트를 임포트
+import CommonLayout from './homepage/components/CommonLayout'; // CommonLayout 임포트
 import '@mantine/core/styles.css';
 
 function App() {
-  return (
-    <MantineProvider withGlobalStyles withNormalizeCSS>
-      <Router>
-        <nav>
-          <Link to="/">Home(나중에)</Link> | <Link to="/search">Search(지우)</Link> | <Link to="/product">Product Detail(겠습니다)</Link>
-        </nav>
+    return (
+        <MantineProvider withGlobalStyles withNormalizeCSS>
+            <Router>
+                <nav>
+                    <Link to="/">Home(나중에)</Link> | <Link to="/search">Search(지우)</Link> | <Link to="/product">Product Detail(겠습니다)</Link> | <Link to="/payment">Payment Page</Link>
+                </nav>
 
-        <Routes>
-          <Route path="/" element={<HomePage />} />
-          <Route path="/search" element={<SearchPage />} />
-          <Route path="/product" element={<ProductDetailPage />} />
-        </Routes>
-      </Router>
-    </MantineProvider>
-  );
+                <Routes>
+                    <Route path="/" element={<CommonLayout><HomePage /></CommonLayout>} />
+                    <Route path="/search" element={<CommonLayout><SearchPage /></CommonLayout>} />
+                    <Route path="/product" element={<CommonLayout><ProductDetailPage /></CommonLayout>} />
+                    <Route path="/payment" element={<PaymentPage />} /> {/* PaymentPage 경로에 layout prop 추가 */}
+                </Routes>
+            </Router>
+        </MantineProvider>
+    );
 }
 
 export default App;
