+++ src/homepage/pages/PaymentPage.jsx
@@ -4,7 +4,7 @@ import { Center, Title } from '@mantine/core';
 const PaymentPage = () => {
     return (
         <Center style={{ height: '100vh', textAlign: 'center' }}>
-            <Text order={1} style={{ fontSize: '48px', color: 'black' }}>결제 페이지</Text>
+            <Title order={1} style={{ fontSize: '48px', color: 'black' }}>결제 페이지</Title>
         </Center>
     );
 };
